Extract session management in tests
