+++ symfony.lock
@@ -718,18 +718,6 @@
     "symfony/intl": {
         "version": "v4.1.3"
     },
-    "symfony/mailer": {
-        "version": "6.1",
-        "recipe": {
-            "repo": "github.com/symfony/recipes",
-            "branch": "main",
-            "version": "4.3",
-            "ref": "97a61eabb351d7f6cb7702039bcfe07fe9d7e03c"
-        },
-        "files": [
-            "config/packages/mailer.yaml"
-        ]
-    },
     "symfony/messenger": {
         "version": "4.3",
         "recipe": {
